Settings and filter show up
